+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to this project will be documented in this file. Format base
 - Implement enumerable editor with radio button
 - Implement `list-box` - a `VBox` emulation of `list-view` and provide consistent styling where used
 - Implement widgets **PlayerSeeker**, **PlayerVolume**, **PlayerLoop** instead of one **PlayerControlsCircle**
+- Implement widget ui controls icons to react on widget size [improves layout]
 - Improve **Tester** widget layout & initial size
 - Improve **StartScreen** styling (no border)
 - Improve `Overlay` show animation transitions